More DAM WIP: refactor controllers and add delete operations.
