+++ modules/admin/conf/institutions.routes
@@ -27,10 +27,13 @@ GET         /:id/reindex                              @controllers.institutions.
 GET         /:id/ingest                               @controllers.institutions.Repositories.ingest(id: String, sync: Boolean ?= false)
 GET         /:id/logo                                 @controllers.institutions.Repositories.updateLogoImage(id: String)
 POST        /:id/logo                                 @controllers.institutions.Repositories.updateLogoImagePost(id: String)
-GET         /:id/data                                 @controllers.institutions.Repositories.uploadData(id: String)
-POST        /:id/data                                 @controllers.institutions.Repositories.uploadDataPost(id: String)
-GET         /:id/data2                                @controllers.institutions.Repositories.uploadDataDirect(id: String)
-POST        /:id/data2                                @controllers.institutions.Repositories.uploadDataDirectPost(id: String, fileName: String ?= "")
-GET         /:id/validate-ead                         @controllers.institutions.Repositories.validateEad(id: String)
-POST        /:id/validate-ead                         @controllers.institutions.Repositories.validateEadPost(id: String)
-GET         /:id/validate-ead2                        @controllers.institutions.Repositories.validateEadFromStorage(id: String)
+
+# Data management
+GET         /:id/data                                 @controllers.institutions.RepositoryData.uploadData(id: String)
+POST        /:id/data                                 @controllers.institutions.RepositoryData.uploadDataPost(id: String)
+GET         /:id/data2                                @controllers.institutions.RepositoryData.uploadDataDirect(id: String)
+POST        /:id/data2                                @controllers.institutions.RepositoryData.uploadDataDirectPost(id: String, fileName: String ?= "")
+POST        /:id/data2/delete                         @controllers.institutions.RepositoryData.deleteDataPost(id: String, fileName: String ?= "")
+GET         /:id/validate-ead                         @controllers.institutions.RepositoryData.validateEad(id: String)
+POST        /:id/validate-ead                         @controllers.institutions.RepositoryData.validateEadPost(id: String)
+GET         /:id/validate-ead2                        @controllers.institutions.RepositoryData.validateEadFromStorage(id: String)
