+++ test/services/storage/MockFileStorage.scala
@@ -36,6 +36,12 @@ case class MockFileStorage(fakeFiles: collection.mutable.ListBuffer[FileStorage#
     Source(fakeFiles
       .filter(p => prefix.forall(p.key.startsWith)).toList)
 
-  override def uri(classifier: String, key: String, duration: FiniteDuration = 10.minutes, forUpload: Boolean = false): URI =
+  override def uri(classifier: String, key: String, duration: FiniteDuration = 10.minutes, contentType: Option[String]): URI =
     new URI(urlPrefix(classifier) + key)
+
+  override def deleteFile(classifier: String, path: String): Future[Unit] = Future {
+    fakeFiles.find(_.key == path).foreach { f =>
+      fakeFiles -= f
+    }
+  }(ec)
 }
